% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_ts.R
\name{arima_ann}
\alias{arima_ann}
\title{Hybrid ARIMA ANN Forecasting Model}
\usage{
arima_ann(y, n, PI = FALSE)
}
\arguments{
\item{y}{A numeric vector or time series}

\item{n}{An integer specifying the forecast horizon}

\item{PI}{A logical flag (default = \code{FALSE}) for generating the prediction interval.}
}
\value{
The forecast of the time series of size \code{n} is generated along with the optional
output of confidence interval (\code{PI} = TRUE) for the forecast.
}
\description{
Hybrid ARIMA ANN Forecasting Model
}
\references{
\itemize{

\item Zhang, G. P. (2003). Time series forecasting using a hybrid ARIMA and neural network model.
Neurocomputing, 50, 159-175.

\item Chakraborty, T., Ghosh, I., Mahajan, T., & Arora, T. (2022). Nowcasting of COVID-19 confirmed
cases: Foundations, trends, and challenges. Modeling, Control and Drug Development for COVID-19
Outbreak Prevention, 1023-1064.}
}
