% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_ts.R
\name{summary_hybridts}
\alias{summary_hybridts}
\title{Summarized score of all the hybrid models implemented in this package}
\usage{
summary_hybridts(train, test)
}
\arguments{
\item{train}{A numeric vector or time series object for training the hybrid models}

\item{test}{A numeric vector or time series object for evaluating the hybrid models}
}
\value{
A data frame where the rows represent the out-of-sample scores for each of
the hybrid models and the columns represent the RMSE, MASE, and sMAPE scores.
}
\description{
Summarized score of all the hybrid models implemented in this package
}
