% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_ts.R
\name{theta_ann}
\alias{theta_ann}
\title{Hybrid Theta ANN Forecasting Model}
\usage{
theta_ann(y, n, PI = FALSE)
}
\arguments{
\item{y}{A numeric vector or time series}

\item{n}{An integer specifying the forecast horizon}

\item{PI}{A logical flag (default = \code{FALSE}) for generating the prediction interval.}
}
\value{
The forecast of the time series of size \code{n} is generated along with the optional
output of confidence interval (\code{PI} = TRUE) for the forecast.
}
\description{
Hybrid Theta ANN Forecasting Model
}
