% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_ts.R
\name{warima_ann}
\alias{warima_ann}
\title{Hybrid WARIMA ANN Forecasting Model}
\usage{
warima_ann(y, n, p = 5, q = 5, PI = FALSE)
}
\arguments{
\item{y}{A numeric vector or time series}

\item{n}{An integer specifying the forecast horizon}

\item{p}{An integer indicating the maximum order of AR process. Default is 5.}

\item{q}{An integer indicating the maximum order of MA process. Default is 5.}

\item{PI}{A logical flag (default = \code{FALSE}) for generating the prediction interval.}
}
\value{
The forecast of the time series of size \code{n} is generated along with the optional
output of confidence interval (\code{PI} = TRUE) for the forecast.
}
\description{
Hybrid WARIMA ANN Forecasting Model
}
\references{
\itemize{
\item Chakraborty, T., Ghosh, I., Mahajan, T., & Arora, T. (2022). Nowcasting of COVID-19 confirmed
cases: Foundations, trends, and challenges. Modeling, Control and Drug Development for COVID-19
Outbreak Prevention, 1023-1064.}
}
